# perf: preallocate in Assign